dubbo-samples-compatible:change groupId and package to org.apache.dubbo
