+++ dubbo-samples-compatible/src/main/resources/META-INF/dubbo/com.alibaba.dubbo.rpc.Filter
@@ -1,1 +1,1 @@
-trace1=com.alibaba.dubbo.samples.basic.TraceFilter
+trace1=TraceFilter
