+++ dubbo-samples-compatible/src/main/resources/spring/dubbo-demo-consumer.xml
@@ -33,6 +33,6 @@
 
     <!-- generate proxy for the remote service, then demoService can be used in the same way as the
     local regular interface -->
-    <dubbo:reference id="demoService" check="false" interface="com.alibaba.dubbo.samples.basic.api.DemoService"/>
+    <dubbo:reference id="demoService" check="false" interface="org.apache.dubbo.samples.basic.api.DemoService"/>
 
 </beans>
